# Jpeg encoder optimization