… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.ScaledCopyTo.cs
… src/ImageSharp/Formats/Jpeg/Components/Block8x8F.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/HuffmanScanDecoder.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/IRawJpegData.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/JpegBlockPostProcessor.cs
… src/ImageSharp/Formats/Jpeg/Components/Decoder/SpectralConverter.cs
… src/ImageSharp/Formats/Jpeg/Components/Encoder/HuffmanLut.cs
… src/ImageSharp/Formats/Jpeg/Components/Encoder/HuffmanScanEncoder.cs
… src/ImageSharp/Formats/Jpeg/Components/FastFloatingPointDCT.Intrinsic.cs
… src/ImageSharp/Formats/Jpeg/Components/FastFloatingPointDCT.cs
… src/ImageSharp/Formats/Jpeg/Components/Quantization.cs
… src/ImageSharp/Formats/Jpeg/Components/ZigZag.Intrinsic.cs
… src/ImageSharp/Formats/Jpeg/Components/ZigZag.cs
… src/ImageSharp/Formats/Jpeg/JpegDecoderCore.cs
… src/ImageSharp/Formats/Jpeg/JpegEncoderCore.cs
… src/ImageSharp/Formats/Tiff/Compression/Decompressors/JpegTiffCompression.cs
… src/ImageSharp/Formats/Tiff/Compression/Decompressors/RgbJpegSpectralConverter.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/BlockOperations/Block8x8F_Quantize.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/BlockOperations/Block8x8F_Transpose.cs
… tests/ImageSharp.Benchmarks/Codecs/Jpeg/EncodeJpeg.cs
… tests/ImageSharp.Benchmarks/Config.HwIntrinsics.cs
… tests/ImageSharp.Benchmarks/Program.cs
… tests/ImageSharp.Tests.ProfilingSandbox/Program.cs
… tests/ImageSharp.Tests/Formats/Jpg/Block8x8FTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/Block8x8Tests.cs
… tests/ImageSharp.Tests/Formats/Jpg/DCTTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/HuffmanScanEncoderTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/QuantizationTests.cs
… tests/ImageSharp.Tests/Formats/Jpg/Utils/JpegFixture.cs
… tests/ImageSharp.Tests/Formats/Jpg/Utils/LibJpegTools.ComponentData.cs