### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

This PR has some performance tweaks & optimizations:

### Quantization
1. Fixed https://github.com/SixLabors/ImageSharp/issues/1751
2. Added SIMD support for zig-zag ordering (encoding only)
3. Quantization now uses reciprocal tables for a multiplication op instead of a division

Benchmark (it's included in the PR):

|   Method |             Job |     Mean |    Error |   StdDev | Ratio |
|--------- |-----------------|---------:|---------:|---------:|------:|
| Quantize | No HwIntrinsics | 73.34 ns | 1.081 ns | 1.011 ns |  1.00 |
| Quantize |             SSE | 24.11 ns | 0.298 ns | 0.279 ns |  0.33 |
| Quantize |             AVX | 15.90 ns | 0.074 ns | 0.065 ns |  0.22 |

### FDCT
1. remade current implementation for ported libjpeg-turbo scalar implementation
2. implemented simd implementation for scalar implementation

Benchmark:
| Method |     Mean |    Error |   StdDev | Ratio |
|--------|---------:|---------:|---------:|------:|
| Master | 36.27 ns | 0.255 ns | 0.226 ns |  1.00 |
| PR     | 30.32 ns | 0.115 ns | 0.108 ns |  0.84 |

### Huffman Encoding
1. Completely redone encoding logic, less if checks & less binary shifts and `&`
2. Small fixes here and there

### Benchmark
It's really hard to test general image encoding/decoding thingy via BenchmarkDotNet so I wrote some custom code for a fixed amount of iterations (300 in following results) of encoding jpeg into **MemoryStream**:
```
// ycbcr 4:4:4
q=100
Master: 26,56ms
PR:     20,22ms
q=90
Master: 19,27ms
PR:     13,95ms
q=75
Master: 18,38ms
PR:     13,45ms
q=50
Master: 16,86ms
PR:     12,56ms

// ycbcr 4:2:0
q=100
Master: 19,41ms
PR:     15,07ms
q=90
Master: 14,56ms
PR:     11,07ms
q=75
Master: 14,05ms
PR:     10,76ms
q=50
Master: 12,67ms
PR:     9,8ms

// luminance only
q=100
Master: 18,59ms
PR:     14,79ms
q=90
Master: 14,62ms
PR:     11,66ms
q=75
Master: 14,23ms
PR:     11,41ms
q=50
Master: 12,98ms
PR:     10,39ms
```