@br3aker 2021-09-12
I guess I need to write some tests (especially for intrinsic paths via remote executor) and maybe some summary comments. But the main code is done and ready for the reviews.

I've actually wanted to take a little break from the project for a week but ETA for 2.0 is almost here afaik so if you want this to be part of the 2.0 - just say so and I'll finish this asap.