@JimBobSquarePants 2021-09-13
>I've actually wanted to take a little break from the project for a week

Take a very well earned break mate! 🙂

 I'll be bumping that deadline along a month. We've still got to figure out the new memory stuff and get WebP reviewed/merged.