@antonfirsov 2021-09-13
Awesome work! I'm quite overloaded with other stuff these days, but will get to this and review in the 2.0 timeframe, which is honestly still 2-3 more months with our current pace I'm afraid.

Is this a draft primarily because of the lack of tests?