@br3aker 2021-09-13
> Is this a draft primarily because of the lack of tests?

I am actually quite surprised that codecov shows positive delta ahah. So it's not that 'drafty' now I guess, There are some things that can be error prone so it's more of a 'please check this **really** carefully before merging' :D