@br3aker 2021-09-25
> If you're happy to have this merged, I'll merge it. Just checking as it's still marked as draft.

Sorry for the late responce. I've actually forgot to mark it as ready to merge. At least Anton should take a look at it too. Don't rush it mates, take your time, this PR won't run away :)

P.S.
Those results though... Glad we could achieve these numbers! Can't wait to see integer based DCT and further decoder optimizations numbers.