@br3aker 2021-09-25
> I wonder if there is a way to reflect this in our tests? Did you notice improvement in any of the encoded images? (Do I get it right that the issue impact dark images mostly?)

It's kind of hard to test as only a very limited range of spectral values were affected - almost `[-0,7..0,7]` range, a bit less. Such low values will be eliminated by any quantization values other than `1` so I doubt that there are any images suffered from this, I don't think that human eye can spot such a tiny detail. New implementation has the same performance for sse/avx path, scalar fallback is faster as it uses only one if check while master branch uses `Vector4.Clamp` which is 2 checks if scalar implementation I believe. But we have the same performance regression as with FDCT - I completely forgot about Vector4 support. Maybe we can also support current implementation via defines?

For testing we can use libjpeg-turbo. As I remember, you're already using it for for spectral checks with approximation of `1f`, new implementation should be a little bit closer to the libjpeg turbo implementation, if not almost equal considering float errors.