@antonfirsov 2021-09-26
> Maybe we can also support current implementation via defines?

`FDCT8x4_sse` seems to be the same as the current implementation, just it's using intrinsic SIMD operations directly. I doubt there is significant perf difference between the two, maybe it's just simpler to use the `Vector4` code if `!Avx.IsSupported && Vector.IsHardwareAccelerated`? (& scalar for the rest.)

Most of today's x86/x64 CPU-s support AVX2 anyways, IMO it's better if the fallback code covers additional platforms instead.