@JimBobSquarePants 2021-09-27
@br3aker Thanks for the excellent explanation. I forgot we were dealing with bits not bytes! Agreed re comments, please add more.