@br3aker 2021-09-28
Hope this passes all tests, fixed almost everything except adding comments so you can review `Vector4` fdct stuff, will add comments tomorrow.

Good news, new scalar transpose implementation is faster than the current one and does not rely on `Vector4` API:

|           Method |             Job |      Mean |     Error |    StdDev | Ratio |
|----------------- |----------------:|----------:|----------:|----------:|------:|
| OLD TransposeInto | No HwIntrinsics | 14.558 ns | 0.0834 ns | 0.0739 ns |  1.00 |
| NEW TransposeInplace | No HwIntrinsics | 12.531 ns | 0.0637 ns | 0.0565 ns |  0.86 |