@antonfirsov 2021-10-02
@JimBobSquarePants I'm not a big fan of the following:
![image](https://user-images.githubusercontent.com/6835152/135713969-54c43cd8-ccba-442d-8d27-93e03e0e4706.png)

Requires us to use admin rights to merge the PR, although there is nothing wrong with the test coverage in reality.